+++ src/index.ts
@@ -7,6 +7,8 @@ export {
   DEFAULT_PAUSED,
   SYMBOLS,
 } from './constants'
-export type { ScrambleText, ScrambleTexts, TextScrambleProps } from './types'
-
+export { createTextScramble, TextScrambleAnimator } from './TextScrambleAnimator'
 export { TextScramble as default, TextScramble }
+
+export type { TextScrambleOptions } from './TextScrambleAnimator'
+export type { ScrambleText, ScrambleTexts, TextScrambleProps } from './types'